requirements.txt
streamlit
pandas
sqlalchemy
pymysql
google-auth
google-api-python-client
python-dotenv

